Moved RatingBracket interface to elo folder
