+++ src/gameplay/elo/types.ts
@@ -1,0 +1,5 @@
+export interface RatingBracket {
+  name: string;
+  min: number;
+  max: number;
+}
